+++ src/components/home/ListProfile.tsx
@@ -1,8 +1,14 @@
-import React, { useEffect } from 'react';
-import * as S from '../../styles/home/ListProfile.styled';
+import React from 'react';
+import { useRecoilValue } from 'recoil';
 import { UserData } from '../../utils/utils';
 import Hashtag from './Hashtag';
 import closeIcon from '../../assets/close-icon.png';
+import * as S from '../../styles/home/ListProfile.styled';
+import { accessTokenState } from '../../atoms';
+import { UserType } from '../../types/ChatType';
+import useGetUserInfo from '../../hooks/useGetUserInfo';
+import useChatAll from '../../hooks/useChatAll';
+import useCreateOrJoinChat from '../../hooks/useCreateOrJoinChat';
 
 function UserProfile({
   userData,
@@ -11,6 +17,20 @@ function UserProfile({
   userData: UserData;
   onCloseModal: () => void;
 }) {
+  const accessToken = useRecoilValue(accessTokenState);
+  const selectedUser: UserType = {
+    id: userData.id || '',
+    name: userData.name || '',
+    picture: userData.image || '',
+  };
+  const userInfo = useGetUserInfo(accessToken);
+  const chatList = useChatAll(accessToken);
+  const handleCreateOrJoinChat = useCreateOrJoinChat(
+    accessToken,
+    chatList,
+    selectedUser,
+    userInfo,
+  );
   const { name, image, language, level, hashtags, intro } = userData;
   let flag;
 
@@ -53,7 +73,9 @@ function UserProfile({
         <S.Right>
           <S.UserInfo>
             <S.UserName>{name}</S.UserName>
-            <S.SendMessage>메시지 보내기</S.SendMessage>
+            <S.SendMessage onClick={handleCreateOrJoinChat}>
+              메시지 보내기
+            </S.SendMessage>
           </S.UserInfo>
           <S.Language>
             <S.Label>관심 언어</S.Label>
